[Refactor] restructure the interfaces of `mainAiCoreService`
